double to float (#18)

+++ tests/src/test_matrix.c
@@ -1,26 +1,32 @@
 
 #include "../include/test_matrix.h"
-    
-double a[] = {
-    1.0, 2.0,
-    3.0, 4.0,
+
+float a[] = {
+    1.0,
+    2.0,
+    3.0,
+    4.0,
 };
 
-double b[] = {
-    4.0, 3.0,
-    2.0, 1.0,
+float b[] = {
+    4.0,
+    3.0,
+    2.0,
+    1.0,
 };
 
 int test_matrix_add()
 {
-    double c[] = {
-        5.0, 5.0,
-        5.0, 5.0,
+    float c[] = {
+        5.0,
+        5.0,
+        5.0,
+        5.0,
     };
 
     Matrix *m1 = matrix_init(2, 2, a);
     Matrix *m2 = matrix_init(2, 2, b);
-    
+
     Matrix *m3 = matrix_add(m1, m2);
     Matrix *expected = matrix_init(2, 2, c);
 
@@ -37,14 +43,16 @@ int test_matrix_add()
 int test_matrix_subtract()
 {
 
-    double c [] = {
-        -3.0, -1.0,
-        1.0, 3.0,
+    float c[] = {
+        -3.0,
+        -1.0,
+        1.0,
+        3.0,
     };
 
     Matrix *m1 = matrix_init(2, 2, a);
     Matrix *m2 = matrix_init(2, 2, b);
-    
+
     Matrix *m3 = matrix_subtract(m1, m2);
     Matrix *expected = matrix_init(2, 2, c);
 
@@ -61,14 +69,16 @@ int test_matrix_subtract()
 int test_matrix_multiply()
 {
 
-    double c[] = {
-        8.0, 5.0,
-        20.0, 13.0,
+    float c[] = {
+        8.0,
+        5.0,
+        20.0,
+        13.0,
     };
 
     Matrix *m1 = matrix_init(2, 2, a);
     Matrix *m2 = matrix_init(2, 2, b);
-    
+
     Matrix *m3 = matrix_multiply(m1, m2);
     Matrix *expected = matrix_init(2, 2, c);
 
@@ -85,13 +95,15 @@ int test_matrix_multiply()
 int test_matrix_multiply_scalar()
 {
 
-    double c[] = {
-        2.0, 4.0,
-        6.0, 8.0,
+    float c[] = {
+        2.0,
+        4.0,
+        6.0,
+        8.0,
     };
 
     Matrix *m1 = matrix_init(2, 2, a);
-    
+
     matrix_scalar_multiply(m1, 2.0);
     Matrix *expected = matrix_init(2, 2, c);
